Fix guidelines and license types on SimpleIcon TS interface
